Update plugin: blog2social.
Name: Blog2Social: Social Media Auto Post & Scheduler
New version: 8.9.0
Previous version: 8.7.1

+++ wp-content/plugins/blog2social/includes/B2S/RePost/Save.php
@@ -1,4 +1,10 @@
 <?php
+if (!defined('ABSPATH')) {
+    exit;
+}
+/**
+ * @phpcs:disable WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedClassFound
+ */
 
 class B2S_RePost_Save {
 
@@ -25,6 +31,7 @@ class B2S_RePost_Save {
     private $linkNoCache;
     private $allowHtml;
     private $default_template;
+    private $setPreFillTextLimit;
 
     function __construct($blogUserId = 0, $b2sUserLang = 'en', $userTimezone = 0, $optionPostFormat = array(), $allowHashTag = true, $bestTimes = array(), $userVersion = 0) {
         $this->userVersion = defined("B2S_PLUGIN_USER_VERSION") ? B2S_PLUGIN_USER_VERSION : (int) $userVersion;
@@ -100,7 +107,7 @@ class B2S_RePost_Save {
                     if ($schedData !== false && is_array($schedData)) {
                         $schedData = array_merge($schedData, $defaultPostData);
                     }
-                    if (in_array($value->networkId, unserialize(B2S_PLUGIN_ALLOW_ADD_LINK)) && isset($this->optionPostFormat[$value->networkId][$value->networkType]['addLink']) && $this->optionPostFormat[$value->networkId][$value->networkType]['addLink'] == false) {
+                    if ( defined('B2S_PLUGIN_ALLOW_ADD_LINK') && in_array($value->networkId, unserialize(B2S_PLUGIN_ALLOW_ADD_LINK)) && isset($this->optionPostFormat[$value->networkId][$value->networkType]['addLink']) && $this->optionPostFormat[$value->networkId][$value->networkType]['addLink'] == false) {
                         if (($value->networkId == 12) || (isset($this->optionPostFormat[$value->networkId][$value->networkType]['format']) && (int) $this->optionPostFormat[$value->networkId][$value->networkType]['format'] == 1)) {
                             $schedData['url'] = '';
                         }
@@ -188,7 +195,7 @@ class B2S_RePost_Save {
             //Share Settings
             $postData['share_settings'] = array('mode' => 0); //share as draft - tiktok
             if ($networkId == 36) {
-                $options = new B2S_Options(B2S_PLUGIN_BLOG_USER_ID);
+                $options = new B2S_Options($this->blogUserId);
                 $optionsShareSettings = $options->_getOption("share_settings");
 
                 //is directly?
@@ -210,6 +217,11 @@ class B2S_RePost_Save {
                 }
                 $postData['custom_title'] = wp_strip_all_tags($this->title);
             }
+
+            //Share as story (form Version 8.8.0 manageable in Post Templates)             
+            if($this->userVersion >= 1 && isset($tempOptionPostFormat[$networkId][$networkType]['share_as_story']) && (int) $tempOptionPostFormat[$networkId][$networkType]['share_as_story'] === 1) {
+                $postData['share_as_story'] = 1;
+            } 
 
             //PostFormat
             if (in_array($networkId, array(1, 2, 3, 12, 17, 19, 24, 43, 44, 45))) {
@@ -263,14 +275,34 @@ class B2S_RePost_Save {
                 if (class_exists('WooCommerce') && function_exists('wc_get_product')) {
                     $wc_product = wc_get_product($this->postId);
                     if ($wc_product != false) {
+
+                        // Current price (already existing)
                         $price = $wc_product->get_price();
-                        if ($price != false && !empty($price)) {
+
+                        if (!empty($price)) {
                             $postData['content'] = stripslashes(preg_replace("/\{PRICE\}/", addcslashes($price, "\\$"), $postData['content']));
+                        }
+
+                        // Regular price
+                        $regular_price = $wc_product->get_regular_price();
+                        if (!empty($regular_price)) {
+                            $postData['content'] = stripslashes(preg_replace("/\{REGULAR_PRICE\}/", addcslashes($regular_price, "\\$"), $postData['content']));
+                        }
+
+                        // Sale price (if any)
+                        $sale_price = $wc_product->get_sale_price();
+                        if (!empty($sale_price)) {
+                            $postData['content'] = stripslashes(preg_replace("/\{SALE_PRICE\}/", addcslashes($sale_price, "\\$"), $postData['content']));
                         }
                     }
                 }
                 $postData['content'] = preg_replace("/\{PRICE\}/", "", $postData['content']);
+                $postData['content'] = preg_replace("/\{REGULAR_PRICE\}/", "", $postData['content']);
+                $postData['content'] = preg_replace("/\{SALE_PRICE\}/", "", $postData['content']);
 
+                //Replace URL from template
+                $postData['content'] = preg_replace("/\{URL\}/", addcslashes($this->url, "\\$"), $postData['content']);
+                
                 $taxonomieReplacements = $hook_filter->get_posting_template_set_taxonomies(array(), $this->postId);
                 if (is_array($taxonomieReplacements) && !empty($taxonomieReplacements)) {
                     foreach ($taxonomieReplacements as $taxonomie => $replacement) {
@@ -410,9 +442,131 @@ class B2S_RePost_Save {
                 }
             }
 
+            // Process comment from template if it exists for all networks
+            if ($this->userVersion >= 2 && B2S_Tools::isCommentAllowed($networkId, $networkType) && isset($tempOptionPostFormat[$networkId][$networkType]['comment']) && !empty($tempOptionPostFormat[$networkId][$networkType]['comment'])) {
+                $postData['comment'] = $this->getCommentByTemplate($tempOptionPostFormat[$networkId][$networkType], $networkId, $networkType);
+            }
+
             return $postData;
         }
         return false;
+    }
+
+    private function getCommentByTemplate($post_template, $networkId, $networkType) {
+        // Get comment text from template
+        $comment = isset($post_template['comment']) && !empty($post_template['comment']) ? $post_template['comment'] : '';
+        
+        if (empty($comment)) {
+            return '';
+        }
+        
+        // Get range settings for comment from short_comment or fallback to short_text
+        $content_min = 0;
+        $content_max = 0;
+        $excerpt_min = 0;
+        $excerpt_max = 0;
+        $limit = 0;
+        
+        if (isset($post_template['short_comment'])) {
+            $content_min = (isset($post_template['short_comment']['range_min'])) ? $post_template['short_comment']['range_min'] : 0;
+            $content_max = (isset($post_template['short_comment']['range_max'])) ? $post_template['short_comment']['range_max'] : 0;
+            $excerpt_min = (isset($post_template['short_comment']['excerpt_range_min'])) ? $post_template['short_comment']['excerpt_range_min'] : 0;
+            $excerpt_max = (isset($post_template['short_comment']['excerpt_range_max'])) ? $post_template['short_comment']['excerpt_range_max'] : 0;
+            $limit = (isset($post_template['short_comment']['limit'])) ? $post_template['short_comment']['limit'] : 0;
+        } else if (isset($this->default_template[$networkId][$networkType]['short_comment'])) {
+            $content_min = (isset($this->default_template[$networkId][$networkType]['short_comment']['range_min'])) ? $this->default_template[$networkId][$networkType]['short_comment']['range_min'] : 0;
+            $content_max = (isset($this->default_template[$networkId][$networkType]['short_comment']['range_max'])) ? $this->default_template[$networkId][$networkType]['short_comment']['range_max'] : 0;
+            $excerpt_min = (isset($this->default_template[$networkId][$networkType]['short_comment']['excerpt_range_min'])) ? $this->default_template[$networkId][$networkType]['short_comment']['excerpt_range_min'] : 0;
+            $excerpt_max = (isset($this->default_template[$networkId][$networkType]['short_comment']['excerpt_range_max'])) ? $this->default_template[$networkId][$networkType]['short_comment']['excerpt_range_max'] : 0;
+            $limit = (isset($this->default_template[$networkId][$networkType]['short_comment']['limit'])) ? $this->default_template[$networkId][$networkType]['short_comment']['limit'] : 0;
+        }
+        
+        // Process placeholders in comment
+        $hook_filter = new B2S_Hook_Filter();
+        
+        // Replace {CONTENT}
+        if (strpos($comment, "{CONTENT}") !== false) {
+            $preContent = addcslashes(B2S_Util::getExcerpt($this->content, (int) $content_min, (int) $content_max), "\\$");
+            $comment = preg_replace("/\{CONTENT\}/", $preContent, $comment);
+        }
+        
+        // Replace {TITLE}
+        if (strpos($comment, "{TITLE}") !== false) {
+            $title = sanitize_text_field($this->title);
+            $comment = preg_replace("/\{TITLE\}/", addcslashes($title, "\\$"), $comment);
+        }
+        
+        // Replace {EXCERPT}
+        if (strpos($comment, "{EXCERPT}") !== false) {
+            $excerpt = (isset($this->excerpt) && !empty($this->excerpt)) ? addcslashes(B2S_Util::getExcerpt($this->excerpt, (int) $excerpt_min, (int) $excerpt_max), "\\$") : '';
+            $comment = preg_replace("/\{EXCERPT\}/", $excerpt, $comment);
+        }
+        
+        // Replace {KEYWORDS}
+        if (strpos($comment, "{KEYWORDS}") !== false) {
+            if ($this->default_template != false && isset($this->default_template[$networkId][$networkType]['disableKeywords']) && $this->default_template[$networkId][$networkType]['disableKeywords'] == true) {
+                $comment = preg_replace("/\{KEYWORDS\}/", '', $comment);
+            } else {
+                $hashtags = ($this->allowHashTag) ? $this->getHashTagsString("", -1, false) : '';
+                $comment = preg_replace("/\{KEYWORDS\}/", addcslashes($hashtags, "\\$"), $comment);
+            }
+        }
+        
+        // Replace {AUTHOR}
+        if (strpos($comment, "{AUTHOR}") !== false) {
+            $authorId = get_post_field('post_author', $this->postId);
+            if (isset($authorId) && !empty($authorId) && (int) $authorId > 0) {
+                $author_name = $hook_filter->get_wp_user_post_author_display_name((int) $authorId);
+                $comment = stripslashes(preg_replace("/\{AUTHOR\}/", addcslashes($author_name, "\\$"), $comment));
+            } else {
+                $comment = preg_replace("/\{AUTHOR\}/", "", $comment);
+            }
+        }
+        
+        // Replace {URL}
+        if (strpos($comment, "{URL}") !== false) {
+            $comment = preg_replace("/\{URL\}/", addcslashes($this->url, "\\$"), $comment);
+        }
+        
+        // Replace WooCommerce price placeholders
+        if (class_exists('WooCommerce') && function_exists('wc_get_product')) {
+            $wc_product = wc_get_product($this->postId);
+            if ($wc_product != false) {
+                $price = $wc_product->get_price();
+                if (!empty($price)) {
+                    $comment = stripslashes(preg_replace("/\{PRICE\}/", addcslashes($price, "\\$"), $comment));
+                }
+                
+                $regular_price = $wc_product->get_regular_price();
+                if (!empty($regular_price)) {
+                    $comment = stripslashes(preg_replace("/\{REGULAR_PRICE\}/", addcslashes($regular_price, "\\$"), $comment));
+                }
+                
+                $sale_price = $wc_product->get_sale_price();
+                if (!empty($sale_price)) {
+                    $comment = stripslashes(preg_replace("/\{SALE_PRICE\}/", addcslashes($sale_price, "\\$"), $comment));
+                }
+            }
+        }
+        
+        $comment = preg_replace("/\{PRICE\}/", "", $comment);
+        $comment = preg_replace("/\{REGULAR_PRICE\}/", "", $comment);
+        $comment = preg_replace("/\{SALE_PRICE\}/", "", $comment);
+        
+        // Replace custom taxonomies
+        $taxonomieReplacements = $hook_filter->get_posting_template_set_taxonomies(array(), $this->postId);
+        if (is_array($taxonomieReplacements) && !empty($taxonomieReplacements)) {
+            foreach ($taxonomieReplacements as $taxonomie => $replacement) {
+                $comment = preg_replace("/\{" . $taxonomie . "\}/", $replacement, $comment);
+            }
+        }
+        
+        // Apply character limit if set
+        if (isset($limit) && (int) $limit > 0) {
+            $comment = B2S_Util::getExcerpt($comment, 0, $limit);
+        }
+        
+        return $comment;
     }
 
     private function getHashTagsString($add = "\n\n", $limit = -1, $shuffle = false) {// limit = -1 => no limit
@@ -477,8 +631,8 @@ class B2S_RePost_Save {
                 'network_type' => (int) $network_type,
                 'network_auth_id' => (int) $network_auth_id,
                 'network_display_name' => $network_display_name,
-                'owner_blog_user_id' => B2S_PLUGIN_BLOG_USER_ID),
-                    array('%d', '%d', '%d', '%s', '%s'));
+                'owner_blog_user_id' => (int) $this->blogUserId),
+                    array('%d', '%d', '%d', '%s', '%d'));
             $networkDetailsId = $wpdb->insert_id;
         }
 
@@ -523,7 +677,7 @@ class B2S_RePost_Save {
         foreach ($schedDataResult as $k => $value) {
             array_push($delete_scheds, $value->b2sPostId);
         }
-        if (!empty($delete_scheds)) {
+        if (!empty($delete_scheds) && defined('B2S_PLUGIN_DIR')) {
             require_once (B2S_PLUGIN_DIR . '/includes/B2S/Post/Tools.php');
             B2S_Post_Tools::deleteUserSchedPost($delete_scheds);
             B2S_Heartbeat::getInstance()->deleteSchedPost();
